requirement.txt
pandas
nltk
emoji
textstat
matplotlib
Pillow
numpy
mtcnn
keras-vggface
keras
tensorflow